tripwhiz/backend/requirements.txt
flask
flask-cors
python-dotenv
googlemaps
requests
gunicorn
